requirements.txt
Django==1.11.7
django-bootstrap3==8.2.3
factory-boy==2.8.1
coverage==4.4.1
coveralls==1.2.0
